Moving files from store to page/journal
