+++ src/effects/journal.effects.ts
@@ -4,7 +4,7 @@ import { Actions, Effect, ofType } from '@ngrx/effects';
 import { switchMap, catchError, map } from 'rxjs/operators';
 import { of } from 'rxjs/observable/of';
 
-import * as journalActions from '../store/journal/journal.actions';
+import * as journalActions from '../pages/journal/journal.actions';
 import { JournalProvider } from '../providers/journal/journal';
 
 @Injectable()
