+++ src/pages/journal/journal.model.ts
