+++ src/pages/journal/journal.reducer.ts
